adding tests and removing dead code
