+++ src/MoBi.UI/Views/EditTableFormulaView.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using MoBi.Presentation.DTO;
 using MoBi.Presentation.Presenter;
 using MoBi.Presentation.Views;
-using OSPSuite.DataBinding;
 using OSPSuite.Presentation.Views;
 using OSPSuite.UI.Controls;
 using OSPSuite.UI.Extensions;
@@ -13,23 +11,14 @@ namespace MoBi.UI.Views
    {
       private IEditTableFormulaPresenter _presenter;
 
-      private readonly ScreenBinder<TableFormulaBuilderDTO> _screenBinder;
-
       public EditTableFormulaView()
       {
          InitializeComponent();
-
-         _screenBinder = new ScreenBinder<TableFormulaBuilderDTO>();
       }
 
       public void AttachPresenter(IEditTableFormulaPresenter presenter)
       {
          _presenter = presenter;
-      }
-
-      public void BindTo(TableFormulaBuilderDTO dtoTableFormulaBuilder)
-      {
-         _screenBinder.BindToSource(dtoTableFormulaBuilder);
       }
 
       public void AddTableView(IView tableView)
